COM-2235 put regroupe attendanceSheet and attendance into AttendanceTable.vue
